+++ composer/autoload_real.php
@@ -13,9 +13,6 @@ class ComposerAutoloaderInit_mediawiki_vendor
         }
     }
 
-    /**
-     * @return \Composer\Autoload\ClassLoader
-     */
     public static function getLoader()
     {
         if (null !== self::$loader) {