+++ wikimedia/parsoid/lib/utils/promise.js
@@ -1,8 +1,0 @@
-/**
- * This module selects Parsoid's default promise implementation.
- * @module
- */
-
-'use strict';
-
-module.exports = require('prfun/wrap')(require('babybird'));