+++ wikimedia/parsoid/src/Config/Env.php
@@ -719,7 +719,7 @@ class Env {
 	/**
 	 * @return array<string,DOMNode[]>
 	 */
-	public function getFragmentMap(): array {
+	public function getDOMFragmentMap(): array {
 		return $this->fragmentMap;
 	}
 
@@ -727,7 +727,7 @@ class Env {
 	 * @param string $id Fragment id
 	 * @return DOMNode[]
 	 */
-	public function getFragment( string $id ): array {
+	public function getDOMFragment( string $id ): array {
 		return $this->fragmentMap[$id];
 	}
 
@@ -736,7 +736,7 @@ class Env {
 	 * @param DOMNode[] $forest DOM forest (contiguous array of DOM trees)
 	 *   to store against the fragment id
 	 */
-	public function setFragment( string $id, array $forest ): void {
+	public function setDOMFragment( string $id, array $forest ): void {
 		$this->fragmentMap[$id] = $forest;
 	}
 
@@ -868,18 +868,20 @@ class Env {
 	/**
 	 * Get an appropriate content handler, given a contentmodel.
 	 *
-	 * @param string|null $forceContentModel An optional content model which
-	 *   will override whatever the source specifies.
+	 * @param string|null &$contentmodel An optional content model which
+	 *   will override whatever the source specifies.  It gets set to the
+	 *   handler which is used.
 	 * @return ContentModelHandler An appropriate content handler
 	 */
 	public function getContentHandler(
-		?string $forceContentModel = null
+		?string &$contentmodel = null
 	): ContentModelHandler {
-		$contentmodel = $forceContentModel ?? $this->pageConfig->getContentModel();
+		$contentmodel = $contentmodel ?? $this->pageConfig->getContentModel();
 		$handler = $this->siteConfig->getContentModelHandler( $contentmodel );
 		if ( !$handler ) {
 			$this->log( 'warn', "Unknown contentmodel $contentmodel" );
-			$handler = $this->siteConfig->getContentModelHandler( 'wikitext' );
+			$contentmodel = 'wikitext';
+			$handler = $this->siteConfig->getContentModelHandler( $contentmodel );
 		}
 		return $handler;
 	}