+++ wikimedia/parsoid/src/Ext/Cite/RefGroup.php
@@ -6,10 +6,8 @@ namespace Wikimedia\Parsoid\Ext\Cite;
 use DOMDocument;
 use DOMElement;
 use stdClass;
-use Wikimedia\Parsoid\Config\Env;
+use Wikimedia\Parsoid\Config\ParsoidExtensionAPI;
 use Wikimedia\Parsoid\Utils\DOMDataUtils;
-use Wikimedia\Parsoid\Utils\DOMUtils;
-use Wikimedia\Parsoid\Utils\Title;
 
 /**
  * Helper class used by `<references>` implementation.
@@ -43,24 +41,21 @@ class RefGroup {
 
 	/**
 	 * Generate leading linkbacks
+	 * @param ParsoidExtensionAPI $extApi
 	 * @param string $href
 	 * @param string|null $group
 	 * @param string $text
 	 * @param DOMDocument $ownerDoc
-	 * @param Env $env
 	 * @return DOMElement
 	 */
 	private static function createLinkback(
-		string $href, ?string $group, string $text, DOMDocument $ownerDoc, Env $env
+		ParsoidExtensionAPI $extApi,
+		string $href, ?string $group, string $text, DOMDocument $ownerDoc
 	): DOMElement {
 		$a = $ownerDoc->createElement( 'a' );
 		$s = $ownerDoc->createElement( 'span' );
 		$textNode = $ownerDoc->createTextNode( $text . ' ' );
-		$title = Title::newFromText(
-			$env->getPageConfig()->getTitle(),
-			$env->getSiteConfig()
-		);
-		$a->setAttribute( 'href', $env->makeLink( $title ) . '#' . $href );
+		$a->setAttribute( 'href', $extApi->getPageUri() . '#' . $href );
 		$s->setAttribute( 'class', 'mw-linkback-text' );
 		if ( $group ) {
 			$a->setAttribute( 'data-mw-group', $group );
@@ -71,11 +66,13 @@ class RefGroup {
 	}
 
 	/**
-	 * @param Env $env
+	 * @param ParsoidExtensionAPI $extApi
 	 * @param DOMElement $refsList
 	 * @param stdClass $ref
 	 */
-	public function renderLine( Env $env, DOMElement $refsList, stdClass $ref ): void {
+	public function renderLine(
+		ParsoidExtensionAPI $extApi, DOMElement $refsList, stdClass $ref
+	): void {
 		$ownerDoc = $refsList->ownerDocument;
 
 		// Generate the li and set ref content first, so the HTML gets parsed.
@@ -83,7 +80,7 @@ class RefGroup {
 		$li = $ownerDoc->createElement( 'li' );
 		$refDir = $ref->dir;
 		$refTarget = $ref->target;
-		$refContent = $ref->content;
+		$refContentId = $ref->contentId;
 		$refGroup = $ref->group;
 		DOMDataUtils::addAttributes( $li, [
 				'about' => '#' . $refTarget,
@@ -99,15 +96,17 @@ class RefGroup {
 				'class' => 'mw-reference-text',
 			]
 		);
-		if ( $refContent ) {
-			$content = $env->getFragment( $refContent )[0];
-			DOMUtils::migrateChildrenBetweenDocs( $content, $reftextSpan );
-			DOMDataUtils::visitAndLoadDataAttribs( $reftextSpan );
+		if ( $refContentId ) {
+			$content = $extApi->getContentDOM( $refContentId );
+			// The data-mw and data-parsoid attributes aren't needed on the ref content
+			// in the references section. The content wrapper will remain in the original
+			// site where the <ref> tag showed up and will retain data-parsoid & data-mw.
+			ParsoidExtensionAPI::migrateChildrenBetweenDocs( $content, $reftextSpan, false );
 		}
 		$li->appendChild( $reftextSpan );
 
 		if ( count( $ref->linkbacks ) === 1 ) {
-			$linkback = self::createLinkback( $ref->id, $refGroup, "↑", $ownerDoc, $env );
+			$linkback = self::createLinkback( $extApi, $ref->id, $refGroup, "↑", $ownerDoc );
 			$linkback->setAttribute( 'rel', 'mw:referencedBy' );
 			$li->insertBefore( $linkback, $reftextSpan );
 		} else {
@@ -118,7 +117,7 @@ class RefGroup {
 
 			foreach ( $ref->linkbacks as $i => $lb ) {
 				$span->appendChild(
-					self::createLinkback( $lb, $refGroup, (string)( $i + 1 ), $ownerDoc, $env )
+					self::createLinkback( $extApi, $lb, $refGroup, (string)( $i + 1 ), $ownerDoc )
 				);
 			}
 		}