+++ wikimedia/parsoid/src/Ext/Cite/RefProcessor.php
@@ -4,7 +4,6 @@ declare( strict_types = 1 );
 namespace Wikimedia\Parsoid\Ext\Cite;
 
 use DOMElement;
-use Wikimedia\Parsoid\Config\Env;
 use Wikimedia\Parsoid\Config\ParsoidExtensionAPI;
 
 /**
@@ -22,18 +21,18 @@ class RefProcessor {
 	}
 
 	/**
+	 * @param ParsoidExtensionAPI $extApi
 	 * @param DOMElement $body
-	 * @param Env $env
 	 * @param array $options
 	 * @param bool $atTopLevel
 	 */
 	public function run(
-		DOMElement $body, Env $env, array $options = [], bool $atTopLevel = false
+		ParsoidExtensionAPI $extApi, DOMElement $body, array $options, bool $atTopLevel
 	): void {
 		if ( $atTopLevel ) {
-			$refsData = new ReferencesData( $env );
+			$refsData = new ReferencesData();
 			References::processRefs( $this->extApi, $refsData, $body );
-			References::insertMissingReferencesIntoDOM( $refsData, $body );
+			References::insertMissingReferencesIntoDOM( $this->extApi, $refsData, $body );
 		}
 	}
 }