+++ wikimedia/parsoid/src/Utils/PipelineUtils.php
@@ -190,6 +190,9 @@ class PipelineUtils {
 	}
 
 	/**
+	 * Convert a DOM node to a token. The node comes from a DOM whose data attributes
+	 * are stored outside the DOM.
+	 *
 	 * @param DOMElement $node
 	 * @param DOMAttr[] $attrs
 	 * @return array
@@ -197,16 +200,19 @@ class PipelineUtils {
 	private static function domAttrsToTagAttrs( DOMElement $node, array $attrs ): array {
 		$out = [];
 		foreach ( $attrs as $a ) {
-			// Not super important since they'll be overwritten by
-			// Parsoid\Wt2Html\PP\Handlers\PrepareDOM
-			if ( !in_array( $a->name, [ 'data-parsoid', DOMDataUtils::DATA_OBJECT_ATTR_NAME ], true ) ) {
+			if ( $a->name !== DOMDataUtils::DATA_OBJECT_ATTR_NAME ) {
 				$out[] = new KV( $a->name, $a->value );
 			}
+		}
+		if ( DOMDataUtils::validDataMw( $node ) ) {
+			$out[] = new KV( 'data-mw', PHPUtils::jsonEncode( DOMDataUtils::getDataMw( $node ) ) );
 		}
 		return [ 'attrs' => $out, 'dataAttrs' => DOMDataUtils::getDataParsoid( $node ) ];
 	}
 
 	/**
+	 * Convert a DOM to tokens. Data attributes for nodes are stored outside the DOM.
+	 *
 	 * @param DOMNode $node The root of the DOM tree to convert to tokens
 	 * @param Token[] $tokBuf This is where the tokens get stored
 	 * @return array
@@ -238,7 +244,7 @@ class PipelineUtils {
 			// getWrapperTokens calls convertDOMToTokens with a DOMElement
 			// and children of dom elements are always text/comment/elements
 			// which are all covered above.
-			Assert::invariant( false, "Should never get here!" );
+			PHPUtils::unreachable( "Should never get here!" );
 		}
 
 		return $tokBuf;
@@ -333,56 +339,47 @@ class PipelineUtils {
 			$wrapperName = $node->nodeName;
 		}
 
-		// Assumed to only be called on nodes that have had data
-		// attributes stored.
 		if ( $node instanceof DOMElement ) {
-			Assert::invariant( !$node->hasAttribute( DOMDataUtils::DATA_OBJECT_ATTR_NAME ),
-				"Expected node to have its data attributes stored" );
-		}
+			Assert::invariant(
+				// No need to look for data-mw as well.
+				// Nodes that have data-mw also have data-parsoid.
+				!$node->hasAttribute( 'data-parsoid' ),
+				"Expected node to have its data attributes loaded" );
 
-		$workNode = null;
-		if ( !( $node instanceof DOMElement ) ) {
-			$workNode = $node->ownerDocument->createElement( $wrapperName );
-		} elseif ( $wrapperName !== $node->nodeName ) {
-			// Create a copy of the node without children
-			$workNode = $node->ownerDocument->createElement( $wrapperName );
-			// Copy over attributes
-			foreach ( DOMCompat::attributes( $node ) as $attribute ) {
-				'@phan-var \DOMAttr $attribute'; // @var \DOMAttr $attribute
-				// "typeof" is ignored since it'll be remove below.
-				// "data-parsoid" will be overwritten with `dataAttribs` when
-				// the token gets to the tree builder, so skip it.  It's
-				// present on the `node` since it has already had its data
-				// attributes stored.
-				if ( !in_array( $attribute->name, [ 'typeof', 'data-parsoid' ], true ) ) {
-					$workNode->setAttribute( $attribute->name, $attribute->value );
+			$nodeData = Util::clone( DOMDataUtils::getNodeData( $node ) );
+
+			if ( $wrapperName !== $node->nodeName ) {
+				// Create a copy of the node without children
+				$workNode = $node->ownerDocument->createElement( $wrapperName );
+
+				// Copy over attributes
+				foreach ( DOMCompat::attributes( $node ) as $attribute ) {
+					'@phan-var \DOMAttr $attribute'; // @var \DOMAttr $attribute
+					// "typeof" is ignored since it'll be remove below.
+					if ( $attribute->name !== 'typeof' ) {
+						$workNode->setAttribute( $attribute->name, $attribute->value );
+					}
+				}
+
+				// We are applying a different wrapper.
+				// So, node's data-parsoid isn't applicable.
+				$nodeData->parsoid = new stdClass;
+			} else {
+				// Shallow clone since we don't want to convert the whole tree to tokens.
+				$workNode = $node->cloneNode( false );
+
+				// Reset 'tsr' since it isn't applicable.
+				// FIXME: The above comment is only true if we are reusing
+				// DOM fragments from cache from previous revisions in
+				// incremental parsing scenarios.
+				if ( isset( $nodeData->parsoid->tsr ) ) {
+					$nodeData->parsoid->tsr = null;
 				}
 			}
-		} else {
-			// Shallow clone since we don't want to convert the whole tree
-			// to tokens.
-			$workNode = $node->cloneNode( false );
 
-			// dataAttribs are not copied over so that we don't inject
-			// broken tsr or dsr values. This also lets these tokens pass
-			// through the sanitizer as stx.html is not set.
-			//
-			// FIXME(arlolra): Presumably, the tsr/dsr portion of the above
-			// comment is with respect to the case where the child nodes are
-			// being dropped.  The stx part though is suspect considering
-			// below where `workNode` is set to the `node` that information
-			// would be preserved.
-			//
-			// We've filed T204279 to look into all this, but for now we'll do
-			// the safe thing and preserve dataAttribs where it makes sense.
-			//
-			// As indicated above, data attributes have already been stored
-			// for `node` so we need to peel them off for the purpose of
-			// cloning.
-			$storedDp = DOMDataUtils::getJSONAttribute( $node, 'data-parsoid', new stdClass );
-			DOMDataUtils::massageLoadedDataParsoid( $storedDp );
-			$storedDp->tsr = null;
-			DOMDataUtils::setDataParsoid( $workNode, $storedDp );
+			DOMDataUtils::setNodeData( $workNode, $nodeData );
+		} else {
+			$workNode = $node->ownerDocument->createElement( $wrapperName );
 		}
 
 		$tokens = self::convertDOMtoTokens( $workNode, [] );
@@ -560,11 +557,8 @@ class PipelineUtils {
 	 * @return array
 	 */
 	public static function makeExpansion( Env $env, array $nodes ): array {
-		foreach ( $nodes as $node ) {
-			DOMDataUtils::visitAndStoreDataAttribs( $node );
-		}
 		$fragmentId = $env->newFragmentId();
-		$env->setFragment( $fragmentId, $nodes );
+		$env->setDOMFragment( $fragmentId, $nodes );
 		return [ 'nodes' => $nodes, 'html' => $fragmentId ];
 	}